third_party/WebKit/LayoutTests/inspector/console/console-functions-expected.txt

CONSOLE MESSAGE: line 22: function simple() {}
CONSOLE MESSAGE: line 23: function simple() {}
CONSOLE MESSAGE: line 22: async function asyncSimple() {}
CONSOLE MESSAGE: line 23: async function asyncSimple() {}
CONSOLE MESSAGE: line 22: function* genSimple() {}
CONSOLE MESSAGE: line 23: function* genSimple() {}
CONSOLE MESSAGE: line 22: function () {}
CONSOLE MESSAGE: line 23: function () {}
CONSOLE MESSAGE: line 22: function (x, y) {}
CONSOLE MESSAGE: line 23: function (x, y) {}
CONSOLE MESSAGE: line 22: function namedArgs(x) {}
CONSOLE MESSAGE: line 23: function namedArgs(x) {}
CONSOLE MESSAGE: line 22: function namedArgs2(x, y) {}
CONSOLE MESSAGE: line 23: function namedArgs2(x, y) {}
CONSOLE MESSAGE: line 22: function ({}) {}
CONSOLE MESSAGE: line 23: function ({}) {}
CONSOLE MESSAGE: line 22: function* whitespace(x) {}
CONSOLE MESSAGE: line 23: function* whitespace(x) {}
CONSOLE MESSAGE: line 22: async function whitespace2(x, y, z) {}
CONSOLE MESSAGE: line 23: async function whitespace2(x, y, z) {}
CONSOLE MESSAGE: line 26: [object Object]
CONSOLE MESSAGE: line 27: [object Object]
Tests that console logging different types of functions correctly.

console-functions.html:22 ƒ simple() {}
console-functions.html:23 ƒ simple()
    arguments: null
    caller: null
    length: 0
    name: "simple"
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:8
    [[Scopes]]: Scopes[1]
console-functions.html:22 async ƒ asyncSimple() {}
console-functions.html:23 async ƒ asyncSimple()
    arguments: (...)
    caller: (...)
    length: 0
    name: "asyncSimple"
    __proto__: AsyncFunction
    [[FunctionLocation]]: console-functions.html:9
    [[Scopes]]: Scopes[1]
console-functions.html:22 ƒ* genSimple() {}
console-functions.html:23 ƒ* genSimple()
    arguments: (...)
    caller: (...)
    length: 0
    name: "genSimple"
    prototype: Generator {}
    __proto__: GeneratorFunction
    [[FunctionLocation]]: console-functions.html:10
    [[IsGenerator]]: true
    [[Scopes]]: Scopes[1]
console-functions.html:22 ƒ () {}
console-functions.html:23 ƒ anonymous()
    arguments: null
    caller: null
    length: 0
    name: ""
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:11
    [[Scopes]]: Scopes[1]
console-functions.html:22 ƒ (x, y) {}
console-functions.html:23 ƒ anonymous(x, y)
    arguments: null
    caller: null
    length: 2
    name: ""
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:12
    [[Scopes]]: Scopes[1]
console-functions.html:22 ƒ namedArgs(x) {}
console-functions.html:23 ƒ namedArgs(x)
    arguments: null
    caller: null
    length: 1
    name: "namedArgs"
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:13
    [[Scopes]]: Scopes[1]
console-functions.html:22 ƒ namedArgs2(x, y) {}
console-functions.html:23 ƒ namedArgs2(x, y)
    arguments: null
    caller: null
    length: 2
    name: "namedArgs2"
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:14
    [[Scopes]]: Scopes[1]
console-functions.html:22 ƒ ({}) {}
console-functions.html:23 ƒ anonymous({})
    arguments: null
    caller: null
    length: 1
    name: ""
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:15
    [[Scopes]]: Scopes[1]
console-functions.html:22 ƒ* whitespace(x) {}
console-functions.html:23 ƒ* whitespace(x)
    arguments: (...)
    caller: (...)
    length: 1
    name: "whitespace"
    prototype: Generator {}
    __proto__: GeneratorFunction
    [[FunctionLocation]]: console-functions.html:16
    [[IsGenerator]]: true
    [[Scopes]]: Scopes[1]
console-functions.html:22 async ƒ whitespace2(x, y, z) {}
console-functions.html:23 async ƒ whitespace2(x, y, z)
    arguments: (...)
    caller: (...)
    length: 3
    name: "whitespace2"
    __proto__: AsyncFunction
    [[FunctionLocation]]: console-functions.html:17
    [[Scopes]]: Scopes[1]
console-functions.html:26 Object
    func0: ƒ simple()
    func1: async ƒ asyncSimple()
    func2: ƒ* genSimple()
    func3: ƒ ()
    func4: ƒ (x, y)
    func5: ƒ namedArgs(x)
    func6: ƒ namedArgs2(x, y)
    func7: ƒ ({})
    func8: ƒ* whitespace(x)
    func9: async ƒ whitespace2(x, y, z)
    __proto__: Object
console-functions.html:27 Object
    func0: ƒ simple()
    func1: async ƒ asyncSimple()
    func2: ƒ* genSimple()
    func3: ƒ ()
    func4: ƒ (x, y)
    func5: ƒ namedArgs(x)
    func6: ƒ namedArgs2(x, y)
    func7: ƒ ({})
    func8: ƒ* whitespace(x)
    func9: async ƒ whitespace2(x, y, z)
    __proto__: Object
